+++ lib/discord/discord-source.test.ts
@@ -83,6 +83,7 @@ const createCtx = (props: CtxProps): FlumeSourceStartContext => ({
   deps: props.deps,
   onStatus: props.onStatus ?? (() => {}),
   reconnect: null,
+  signal: undefined,
 })
 
 const simulateReadySequence = () => {