+++ lib/slack/slack-source.test.ts
@@ -105,6 +105,7 @@ const createCtx = (props: CtxProps): FlumeSourceStartContext => ({
   deps: props.deps,
   onStatus: props.onStatus ?? (() => {}),
   reconnect: null,
+  signal: undefined,
 })
 
 describe("FlumeSlackSource", () => {